Add clean command to only remove dependencies but not top level folder

+++ gitman/commands.py
@@ -209,13 +209,14 @@ def lock(*names, root=None):
 
 
 @restore_cwd
-def delete(*, root=None, force=False):
+def delete(*, root=None, force=False, keep_location=False):
     """Delete dependencies for a project.
 
     Optional arguments:
 
     - `root`: specifies the path to the root working tree
     - `force`: indicates uncommitted changes can be overwritten
+    - `keep_location`: delete top level folder or keep the location
 
     """
     log.info("Deleting dependencies...")
@@ -232,7 +233,10 @@ def delete(*, root=None, force=False):
         common.dedent(level=0)
         common.show("Deleting all dependencies...", color='message', log=False)
         common.newline()
-        config.uninstall_dependencies()
+        if keep_location:
+            config.clean_dependencies()
+        else:
+            config.uninstall_dependencies()
 
     return _display_result("delete", "Deleted", count, allow_zero=True)
